Moving all array-like data classes into the same package
