+++ engine/runtime/src/main/java/org/enso/interpreter/runtime/data/EnsoFile.java
@@ -15,6 +15,7 @@ import java.util.function.IntFunction;
 
 import org.enso.interpreter.dsl.Builtin;
 import org.enso.interpreter.runtime.EnsoContext;
+import org.enso.interpreter.runtime.data.vector.ArrayOverBuffer;
 import org.enso.interpreter.runtime.error.PanicException;
 import org.enso.interpreter.runtime.library.dispatch.TypesLibrary;
