+++ engine/runtime/src/main/java/org/enso/interpreter/runtime/scope/TopLevelScope.java
@@ -20,7 +20,7 @@ import org.enso.interpreter.EnsoLanguage;
 import org.enso.interpreter.runtime.EnsoContext;
 import org.enso.interpreter.runtime.Module;
 import org.enso.interpreter.runtime.builtin.Builtins;
-import org.enso.interpreter.runtime.data.Array;
+import org.enso.interpreter.runtime.data.vector.Array;
 import org.enso.interpreter.runtime.type.Types;
 import org.enso.interpreter.util.ScalaConversions;
 import org.enso.pkg.Package;
